Fixed almost all warnings

+++ src/main/java/com/cerbon/queen_bee/client/sound/QueenBeeSoundInstance.java
@@ -17,9 +17,9 @@ public abstract class QueenBeeSoundInstance extends AbstractTickableSoundInstanc
     protected QueenBeeSoundInstance(QueenBeeEntity queenBee, SoundEvent pSoundEvent, SoundSource pSource) {
         super(pSoundEvent, pSource, SoundInstance.createUnseededRandom());
         this.queenBee = queenBee;
-        this.x = (double)((float)queenBee.getX());
-        this.y = (double)((float)queenBee.getY());
-        this.z = (double)((float)queenBee.getZ());
+        this.x = ((float)queenBee.getX());
+        this.y = ((float)queenBee.getY());
+        this.z = ((float)queenBee.getZ());
         this.looping = true;
         this.delay = 0;
         this.volume = 0.0F;
@@ -34,9 +34,9 @@ public abstract class QueenBeeSoundInstance extends AbstractTickableSoundInstanc
         }
 
         if (!this.queenBee.isRemoved() && !this.hasSwitched) {
-            this.x = (double)((float)this.queenBee.getX());
-            this.y = (double)((float)this.queenBee.getY());
-            this.z = (double)((float)this.queenBee.getZ());
+            this.x = ((float)this.queenBee.getX());
+            this.y = ((float)this.queenBee.getY());
+            this.z = ((float)this.queenBee.getZ());
             float f = (float)this.queenBee.getDeltaMovement().horizontalDistance();
             if (f >= 0.01F) {
                 this.pitch = Mth.lerp(Mth.clamp(f, 0.7F, 1.0F), 0.7F, 1.0F);
